// webapp/src/app/types/product.ts
export type Product = {
    name: string
    description: string | null
    html_url: string
    owner: string
    avatar_url: string | null
    language: string | null
    stargazers_count: number
    num_code_files: number | null
    repo_created_at: string
    repo_pushed_at: string
    created_at: string
}